Refactor Forms's file & folder structure
