--- src/components/common/Form/styles.scss
@@ -1,31 +1,0 @@
-// End
-.end {
-  margin-bottom: 0;
-}
-
-// Headlines
-.headline {
-  font: {
-    size: 1.5rem;
-    weight: 300;
-  }
-  margin-bottom: 1.75rem;
-}
-
-// Subheadlines
-.subheadline {
-  font: {
-    size: 1.1rem;
-    weight: 300;
-  }
-  margin-bottom: 1.25rem;
-}
-
-// Meta context
-.meta {
-  padding-top: 0.25rem;
-
-  .context {
-    font-size: 1rem;
-  }
-}
