+++ src/components/common/Forms/FormGroup/styles.scss
@@ -1,0 +1,4 @@
+// End
+.end {
+  margin-bottom: 0;
+}
