+++ src/components/common/Forms/FormHL/styles.scss
@@ -1,0 +1,8 @@
+// Headlines
+.headline {
+  font: {
+    size: 1.5rem;
+    weight: 300;
+  }
+  margin-bottom: 1.75rem;
+}
