+++ src/components/common/Forms/FormMeta/styles.scss
@@ -1,0 +1,8 @@
+// Meta context
+.meta {
+  padding-top: 0.25rem;
+
+  .context {
+    font-size: 1rem;
+  }
+}
