+++ src/components/common/Forms/FormSHL/styles.scss
@@ -1,0 +1,8 @@
+// Subheadlines
+.subheadline {
+  font: {
+    size: 1.1rem;
+    weight: 300;
+  }
+  margin-bottom: 1.25rem;
+}
